refactor: ConsoleApplication으로 시작점 변경
