+++ src/main/java/com/devcourse/springbootbasic/SpringbootBasicApplication.java
@@ -1,6 +1,6 @@
 package com.devcourse.springbootbasic;
 
-import com.devcourse.springbootbasic.application.controller.PlatformController;
+import com.devcourse.springbootbasic.application.ConsoleApplication;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
@@ -9,7 +9,7 @@ public class SpringbootBasicApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(SpringbootBasicApplication.class, args)
-                .getBean(PlatformController.class)
+                .getBean(ConsoleApplication.class)
                 .run();
     }
